Added a sendPacket function

+++ protocol.c
@@ -178,8 +178,8 @@ int senderAcknowledge(field_t *sendPacket, field_t *receivedPacket, digit_t *sen
     return ENC_ACCEPT_PACKET;
 }
 
-int increaseCounter(long *counter) {
-	long nextValue = *counter + 1;
+int increaseCounter(uint32_t *counter) {
+	uint32_t nextValue = *counter + 1;
 	if(*counter > nextValue) {
 		*counter = nextValue;
 		return ENC_COUNTER_WRAPAROUND;
